[interpreter] Rename module-interpreter to source-level-interpreter
## Summary

To make it clear that the interpreter runs on source-level code.

## Test Plan

👀
